// src/components/weather/components/types.ts
import { BoxProps, CardProps, StackProps } from "@mui/material";
export interface ISearch {
  city: string;
  setCity: (x: string) => void;
  handlerSearchCity: () => void;
  loading: boolean;
  error: string;
}

export interface IWeatherResponse {
  location: {
    name: string;
    region: string;
    country: string;
    lat: number;
    lon: number;
    tz_id: string;
    localtime_epoch: number;
    localtime: string;
  };
  current: {
    last_updated_epoch: number;
    last_updated: string;
    temp_c: number;
    temp_f: number;
    is_day: number;
    condition: {
      text: string;
      icon: string;
      code: number;
    };
    wind_mph: number;
    wind_kph: number;
    wind_degree: number;
    wind_dir: string;
    pressure_mb: number;
    pressure_in: number;
    precip_mm: number;
    precip_in: number;
    humidity: number;
    cloud: number;
    feelslike_c: number;
    feelslike_f: number;
    windchill_c: number;
    windchill_f: number;
    heatindex_c: number;
    heatindex_f: number;
    dewpoint_c: number;
    dewpoint_f: number;
    vis_km: number;
    vis_miles: number;
    uv: number;
    gust_mph: number;
    gust_kph: number;
  };
}

export interface ICurrentWeather extends Omit<CardProps, "weather"> {
  weather: IWeatherResponse;
}

export interface ICardWeather extends Omit<StackProps, "weather"> {
  weather: IWeatherResponse;
}

export type TWeatherCondition = {
  code: number;
  name: string;
  icon: JSX.Element;
};

export interface IDetailsWeather extends Omit<BoxProps, "weather"> {
  weather: IWeatherResponse;
}